refactor: refactor scheduler service to separate client
