+++ batch/src/main/resources/application-dev.yml
@@ -17,7 +17,8 @@ schedules:
   cron:
     dividend: "0 0 0 * * *"
 
-fmp:
-  key: ${FMP_API_KEY}
-  base-url: "https://financialmodelingprep.com/api/v3"
-  stock-dividend-calendar-postfix: "/stock_dividend_calendar"
+financial:
+  fmp:
+    key: ${FMP_API_KEY}
+    base-url: "https://financialmodelingprep.com"
+    stock-dividend-calendar-postfix: "/api/v3/stock_dividend_calendar"
